+++ api/src/main/java/org/infinispan/api/store/WriteResult.java
@@ -1,4 +1,4 @@
-package org.infinispan.api.reactive;
+package org.infinispan.api.store;
 
 import org.infinispan.api.Experimental;
