+++ api/src/main/java/org/infinispan/api/store/listener/KeyValueStoreListener.java
@@ -1,9 +1,9 @@
-package org.infinispan.api.reactive.listener;
+package org.infinispan.api.store.listener;
 
 import org.infinispan.api.Experimental;
 
 /**
- * KeyValueStoreListener interface to be used with {@link org.infinispan.api.reactive.KeyValueStore}
+ * KeyValueStoreListener interface to be used with {@link org.infinispan.api.store.KeyValueStore}
  * Listeners, Client or Embedded, must implement this interface
  *
  * @since 10.0
