+++ client/infinispan-key-value-store-hotrod/src/main/java/org/infinispan/api/client/impl/InfinispanClientImpl.java
@@ -12,11 +12,11 @@ import java.util.concurrent.ExecutorService;
 import org.infinispan.api.configuration.ClientConfig;
 import org.infinispan.api.Infinispan;
 import org.infinispan.api.exception.InfinispanConfigurationException;
-import org.infinispan.api.reactive.client.impl.KeyValueStoreImpl;
+import org.infinispan.api.store.client.impl.KeyValueStoreImpl;
 import org.infinispan.api.exception.InfinispanException;
 import org.infinispan.api.marshalling.Marshaller;
-import org.infinispan.api.reactive.KeyValueStore;
-import org.infinispan.api.reactive.KeyValueStoreConfig;
+import org.infinispan.api.store.KeyValueStore;
+import org.infinispan.api.store.KeyValueStoreConfig;
 import org.infinispan.client.hotrod.RemoteCache;
 import org.infinispan.client.hotrod.RemoteCacheManager;
 import org.infinispan.client.hotrod.marshall.ProtoStreamMarshaller;
