+++ client/infinispan-key-value-store-hotrod/src/test/java/org/infinispan/api/client/impl/KeyValueStoreTest.java
@@ -12,9 +12,9 @@ import java.util.AbstractMap;
 import org.infinispan.api.Infinispan;
 import org.infinispan.api.InfinispanClient;
 import org.infinispan.api.configuration.ClientConfig;
-import org.infinispan.api.reactive.KeyValueStore;
-import org.infinispan.api.reactive.KeyValueStoreConfig;
-import org.infinispan.api.reactive.WriteResult;
+import org.infinispan.api.store.KeyValueStore;
+import org.infinispan.api.store.KeyValueStoreConfig;
+import org.infinispan.api.store.WriteResult;
 import org.infinispan.client.hotrod.configuration.ConfigurationBuilder;
 import org.infinispan.client.hotrod.test.HotRodClientTestingUtil;
 import org.infinispan.client.hotrod.test.SingleHotRodServerTest;
