+++ client/infinispan-key-value-store-hotrod/src/test/java/org/infinispan/api/client/impl/SearchKeyValueStoreTest.java
@@ -16,11 +16,11 @@ import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
 import org.infinispan.api.Infinispan;
-import org.infinispan.api.reactive.KeyValueEntry;
-import org.infinispan.api.reactive.KeyValueStore;
-import org.infinispan.api.reactive.KeyValueStoreConfig;
-import org.infinispan.api.reactive.query.ContinuousQueryRequestBuilder;
-import org.infinispan.api.reactive.query.QueryRequest;
+import org.infinispan.api.store.KeyValueEntry;
+import org.infinispan.api.store.KeyValueStore;
+import org.infinispan.api.store.KeyValueStoreConfig;
+import org.infinispan.api.store.query.ContinuousQueryRequestBuilder;
+import org.infinispan.api.store.query.QueryRequest;
 import org.infinispan.client.hotrod.RemoteCacheManager;
 import org.infinispan.client.hotrod.test.HotRodClientTestingUtil;
 import org.infinispan.client.hotrod.test.InternalRemoteCacheManager;
