+++ integrationtests/security-it/src/test/java/org/infinispan/test/integration/security/utils/Deployments.java
@@ -44,7 +44,7 @@ public final class Deployments {
                   new File("target/test-libs/infinispan-commons.jar"),
                   new File("target/test-libs/caffeine.jar"),
                   new File("target/test-libs/rxjava.jar"),
-                  new File("target/test-libs/reactive-streams.jar"),
+                  new File("target/test-libs/store-streams.jar"),
                   new File("target/test-libs/jboss-marshalling.jar"),
                   new File("target/test-libs/jboss-marshalling-river.jar"),
                   new File("target/test-libs/protostream.jar"),
